(BSR)[PRO] refactor: move link venues in a file
